Add configurations and email confirmation. -- Added email confirmation with web controller. -- Add confirmation controllers. -- Removed unnecessary updating. -- Added pattern updates and email, ranges too. -- Add creation and update validations.

+++ src/main/java/com/g2/personalaccount/dto/responses/AccountResponse.java
@@ -14,6 +14,6 @@ public class AccountResponse {
   private String holderFirstName;
   private String holderLastName;
   private String email;
-  private Integer ssn;
-  private Integer voterCardId;
+  private Long ssn;
+  private Long voterCardId;
 }
